Working lasers and dots components for board
