+++ src/components/Board/index.vue
@@ -1,28 +1,10 @@
 <template>
   <svg ref="grid" class="grid" :width="totalWidth" :height="totalHeight">
     <!-- DOTS -->
-    <g v-for="(row, y) in level.grid.rows" :key="y">
-      <g v-for="(column, x) in level.grid.cols" :key="x">
-        <circle :cx="x * tileSize" :cy="y * tileSize" r="1" fill="#edeaf4" />
-      </g>
-    </g>
+    <board-dots />
 
     <!-- LASER PATH -->
-    <g
-      v-for="(laser, index) in individualLaserPath"
-      :key="'laser' + index"
-      :v-if="individualLaserPath.length > 0"
-      class="lasers"
-    >
-      <path
-        :d="laser"
-        stroke-dasharray="8 8"
-        fill="transparent"
-        stroke="red"
-        stroke-width="3"
-        class="laserPath"
-      />
-    </g>
+    <board-lasers />
 
     <!-- CELLS -->
     <app-cell
@@ -43,6 +25,7 @@
     >
       <app-photon
         name
+        :particle="particle"
         :intensity="particle.intensity"
         :are="particle.a.re"
         :aim="particle.a.im"
@@ -75,6 +58,8 @@ import Grid from '@/engine/Grid';
 import Level from '@/engine/Level';
 import { ParticleInterface, CellInterface, HintInterface } from '@/engine/interfaces';
 import AppCell from '@/components/Board/AppCell.vue';
+import BoardLasers from '@/components/Board/BoardLasers.vue';
+import BoardDots from '@/components/Board/BoardDots.vue';
 import AppPhoton from '@/components/AppPhoton.vue';
 import SpeechBubble from '@/components/SpeechBubble.vue';
 
@@ -82,6 +67,8 @@ import SpeechBubble from '@/components/SpeechBubble.vue';
   components: {
     AppCell,
     AppPhoton,
+    BoardLasers,
+    BoardDots,
     SpeechBubble
   }
 })
@@ -108,15 +95,19 @@ export default class Board extends Vue {
     this.tileSize = 64;
   }
 
-  get lasers(): ParticleInterface[] {
-    return this.level.grid.computePaths();
-  }
-
   get totalWidth(): number {
     return this.level.grid.cols * this.tileSize;
   }
   get totalHeight(): number {
     return this.level.grid.rows * this.tileSize;
+  }
+
+  /**
+   * Compute the cell center at a specific coordinate for grid dots
+   * @returns x, y pixel coordinates
+   */
+  centerCoord(val: number): number {
+    return (val + 0.5) * this.tileSize;
   }
 
   computeParticleStyle(particle: ParticleInterface): {} {
@@ -131,14 +122,6 @@ export default class Board extends Vue {
   }
 
   /**
-   * Compute the cell center at a specific coordinate for grid dots
-   * @returns x, y pixel coordinates
-   */
-  centerCoord(val: number): number {
-    return (val + 0.5) * this.tileSize;
-  }
-
-  /**
    * Used to move or swap cells
    * @params coord to move to
    * @returns boolean
@@ -148,60 +131,10 @@ export default class Board extends Vue {
     const targetCell = this.level.grid.get(coord);
     const mutatedCells: Cell[] = this.level.grid.move(sourceCell, targetCell);
     mutatedCells.forEach((cell) => {
-      // this.mutationUpdateGridCell(cell);
       this.level.grid.set(cell);
     });
     this.$emit('updateSimulation');
-  }
-
-  /**
-   * Create laser path through the lasers points
-   * @returns SVG laser path
-   */
-  laserPath(): string {
-    let pathStr = '';
-    if (this.lasers.length > 0) {
-      const originX = this.centerCoord(this.lasers[0].coord.x);
-      const originY = this.centerCoord(this.lasers[0].coord.y);
-      pathStr += `M ${originX} ${originY} `;
-      this.lasers.forEach((laser: any) => {
-        const x = this.centerCoord(laser.coord.x);
-        const y = this.centerCoord(laser.coord.y);
-        pathStr += ` L ${x} ${y} `;
-      });
-      pathStr += ' ';
-    }
-    return pathStr;
-  }
-
-  get individualLaserPath(): string[] {
-    const pathsStr: string[] = [];
-    if (this.lasers.length > 0) {
-      this.lasers.forEach((laser: any) => {
-        let pathStr = '';
-        const originX = this.centerCoord(laser.coord.x);
-        const originY = this.centerCoord(laser.coord.y);
-        pathStr += `M ${originX} ${originY} `;
-        switch (laser.direction) {
-          case 0:
-            pathStr += ` H ${this.centerCoord(laser.coord.x + 1)}`;
-            break;
-          case 90:
-            pathStr += ` V ${this.centerCoord(laser.coord.y - 1)}`;
-            break;
-          case 180:
-            pathStr += ` H ${this.centerCoord(laser.coord.x - 1)}`;
-            break;
-          case 270:
-            pathStr += ` V ${this.centerCoord(laser.coord.y + 1)}`;
-            break;
-          default:
-            throw new Error(`Laser has wrong direction: ${laser.direction}°`);
-        }
-        pathsStr.push(pathStr);
-      });
-    }
-    return pathsStr;
+    this.$emit('updateGrid');
   }
 
   /**
@@ -214,11 +147,6 @@ export default class Board extends Vue {
       const originX = this.centerCoord(this.particles[0].coord.x);
       const originY = this.centerCoord(this.particles[0].coord.y);
       pathStr += `M ${originX} ${originY} `;
-      this.lasers.forEach((laser: any) => {
-        const x = this.centerCoord(laser.coord.x);
-        const y = this.centerCoord(laser.coord.y);
-        pathStr += ` L ${x} ${y} `;
-      });
     }
     return pathStr;
   }
